chore: Clean up the code
Remove some of code that was redundant or not required.

This includes:
	- Unused css
	- Image (will be added later)

Move the html code around:
	- Move `<header>` from `<head>`
	- Enclose the main text in `<main>`
	- Use `<footer>` instead of `<div>`
	- Remove some text from the page

Change the styling:
	- Use flex instead of absolute tag
	- Utilize the `.active` class
	- Don't circle the social icons
	- Style the <hr> tag

Add a footer:
	- The footer has the following socials:
		- Twitter (X?)
		- LinkedIn (link not yet added)
		- Github
	- Push the footer to the bottom (expected)

Make the site mobile friendly.
Instead of using `@media-query`, `flex` is used.
This reduces the amount of code required and makes the codebase clean.

(
The code from now on is attributed to the respective contributors.
Any previous copyright claim is revoked from here on in favour of
per individual contributors.
)

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>