chore: Add a tagline
The homepage now has a tagline as per the original design.
A tagline/quote is added (temporary for now).

Mobile responsiveness has been enhanced.
Now viewing the page in mobile is atleast pleasant.

There is still an issue on how the image is displayed, only the
center part is actually viewable.
This gives different users different views.
Some see a beautiful background, others not.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>