chore: Rewrite homepage to be more appealing
Write the homepage with styling in mind.
Now it supports both light and dark mode internally.

Themes supported are dark and light theme (obviously),
internally css `@media (prefers-color-scheme: dark|light)`
is used. External methods like a button may be introduced
later.

The homepage now features:
    - Two images (not responsive enough though)
    - Featured project
    - Footer

Use of colours (or colors) variation is kept to the minimum.
This is to accomodate the Blue|White|Black|Grey colour choices
made ealier.
Strength of the colour have been reduced from their "native"
values for more comfy reading of text.
No gurantee is made that it is the case.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>